Functionality to edit the direction description based on item availability
if you pickup an item from a room from a certain direction, if you look toward that direction it will not tell you that you have an item there. if you drop an item, it will add a message saying what is the item dropped in the set direction for that item in the specific room the item supposed to be at.